refactor: Post Page 용 컴포넌트 관리폴더 이름을 sections -> components로 변경, 각 파일 외부참조 경로 절대참조로 수정
